refactor: 프로젝트 구조 변경
util, http, dispatcher 패키지에 있는 클래스들을 재배치했다. http 패키지 내부도 하위 패키지로 나누기보단 클래스들을 나열해두어 복잡도를 낮추었다. util 내부의 클래스는 주로 쓰는 클래스 근처에 배치해두었다. Dispatcher 역시 Handler가 보내준 request를 적절한 method나 정적 파일로 라우팅해주는 역할이라서 Handler가 있는 패키지에 배치하였다.
