+++ src/main/java/webserver/http/RequestBody.java
@@ -1,4 +1,4 @@
-package http.body;
+package webserver.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
