+++ src/main/java/webserver/http/SessionManager.java
@@ -1,4 +1,4 @@
-package http;
+package webserver.http;
 
 import model.User;
